+++ src/devices/zen.ts
@@ -23,7 +23,7 @@ const definitions: Definition[] = [
         ota: ota.zigbeeOTA,
         exposes: [e.climate().withSetpoint('occupied_heating_setpoint', 10, 30, 0.5)
             .withSetpoint('occupied_cooling_setpoint', 10, 31, 0.5).withLocalTemperature()
-            .withSystemMode(['off', 'auto', 'heat', 'cool']).withRunningState(['idle', 'heat', 'cool'])
+            .withSystemMode(['off', 'auto', 'heat', 'cool', 'emergency_heating']).withRunningState(['idle', 'heat', 'cool'])
             .withLocalTemperatureCalibration().withFanMode(['auto', 'on'])],
         configure: async (device, coordinatorEndpoint, logger) => {
             const endpoint = device.getEndpoint(3) || device.getEndpoint(1);