+++ src/lib/generateDefinition.ts
@@ -139,10 +139,11 @@ export async function generateDefinition(device: Zh.Device): Promise<{externalDe
     // It is possible to better check if device should be considered multiEndpoint
     // based, for example, on generator arguments(i.e. presence of "endpointNames"),
     // but this will be enough for now.
-    const multiEndpoint = device.endpoints.length > 1;
+    const endpointsWithoutGreenPower = device.endpoints.filter((e) => e.ID !== 242);
+    const multiEndpoint = endpointsWithoutGreenPower.length > 1;
     if (multiEndpoint) {
         const endpoints: {[n: string]: number} = {};
-        for (const endpoint of device.endpoints) {
+        for (const endpoint of endpointsWithoutGreenPower) {
             endpoints[endpoint.ID.toString()] = endpoint.ID;
         }
         // Add to beginning for better visibility.
@@ -231,7 +232,6 @@ const inputExtenders: Extender[] = [
 ];
 
 const outputExtenders: Extender[] = [
-    [['genOta'], async (d, eps) => [new Generator({extend: m.ota, source: 'ota'})]],
     [['genOnOff'], async (d, eps) => [
         new Generator({extend: m.commandsOnOff, args: maybeEndpointArgs(d, eps), source: 'commandsOnOff'}),
     ]],